estore/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
sqlparse==0.5.3
